refactor: surface artist radio as a dedicated header button
Move the "radio" action out of the artist options (⋯) menu into a
dedicated button in the artist header actions bar, next to shuffle,
matching Navidrome's own artist page layout.

The radio logic is extracted into a reusable useArtistRadio hook and the
button (with a new Actions.RadioIcon) is only rendered when the server
advertises the sonicSimilarity extension. i18n keys moved under
artist.buttons.radio / artist.radio.empty.
