+++ src/app/components/options/buttons.tsx
@@ -9,7 +9,6 @@ import {
   PlayIcon,
   PlusIcon,
   PodcastIcon,
-  Radio,
   Trash,
 } from 'lucide-react'
 import { useTranslation } from 'react-i18next'
@@ -58,19 +57,6 @@ function PlayLast({ variant = 'dropdown', ...props }: DropdownMenuItemProps) {
       variant={variant}
       icon={<ListEnd className="mr-2 h-4 w-4" />}
       label={t('options.addLast')}
-      {...props}
-    />
-  )
-}
-
-function StartRadio({ variant = 'dropdown', ...props }: DropdownMenuItemProps) {
-  const { t } = useTranslation()
-
-  return (
-    <MenuItemFactory
-      variant={variant}
-      icon={<Radio className="mr-2 h-4 w-4" />}
-      label={t('options.startRadio')}
       {...props}
     />
   )
@@ -213,7 +199,6 @@ export const OptionsButtons = {
   Play,
   PlayNext,
   PlayLast,
-  StartRadio,
   Download,
   AddToPlaylistOption,
   EditPlaylist,
