+++ src/i18n/locales/pt.json
@@ -50,7 +50,11 @@
     "buttons": {
       "shuffle": "Tocar a rádio de {{artist}} no modo aleatório",
       "options": "Mais opções para {{artist}}",
-      "play": "Tocar a rádio de {{artist}}"
+      "play": "Tocar a rádio de {{artist}}",
+      "radio": "Iniciar rádio baseada em {{artist}}"
+    },
+    "radio": {
+      "empty": "Não foram encontradas faixas semelhantes para iniciar uma rádio."
     },
     "info": {
       "albumsCount_one": "{{count}} álbum",
@@ -655,8 +659,6 @@
     "play": "Reproduzir",
     "playNext": "Tocar a seguir",
     "addLast": "Adicionar ao final",
-    "startRadio": "Iniciar rádio",
-    "startRadioEmpty": "Não foram encontradas faixas semelhantes para iniciar uma rádio.",
     "download": "Descarregar",
     "playlist": {
       "add": "Adicionar à playlist",
